+++ crates/topcoat-core/src/context.rs
@@ -1,6 +1,7 @@
 mod binding;
 mod context_map;
 mod id;
+pub mod test;
 
 use std::{
     any::{Any, TypeId},
@@ -13,6 +14,7 @@ use std::{
 use bit_set::BitSet;
 pub use context_map::*;
 pub use id::*;
+pub use test::*;
 
 use crate::{abort::AbortStore, memoize::MemoizeCache};
 
@@ -312,46 +314,6 @@ mod private {
     pub trait ContextValues {
         fn assert_unique_types();
         fn install(self, cx: &mut Cx);
-    }
-}
-
-/// Assembles a [`Cx`] from scratch for tests.
-#[derive(Debug, Default)]
-pub struct CxTestBuilder {
-    cx: Cx,
-}
-
-impl CxTestBuilder {
-    /// Creates an empty builder.
-    #[must_use]
-    pub fn new() -> Self {
-        Self::default()
-    }
-
-    /// Registers `value` on the app context.
-    #[must_use]
-    pub fn app_context<T>(mut self, value: T) -> Self
-    where
-        T: Any + Send + Sync,
-    {
-        let _ = self.cx.app_context_mut().insert(value);
-        self
-    }
-
-    /// Registers `value` on the request root context.
-    #[must_use]
-    pub fn request_context<T>(mut self, value: T) -> Self
-    where
-        T: Any + Send + Sync,
-    {
-        let _ = self.cx.insert(value);
-        self
-    }
-
-    /// Consumes the builder, returning the assembled [`Cx`].
-    #[must_use]
-    pub fn build(self) -> Cx {
-        self.cx
     }
 }
