+++ crates/topcoat/docs/mail.md
@@ -121,7 +121,7 @@ let from_url = SmtpTransport::from_url("smtps://user:pass@smtp.example.com:465")
 
 ```rust
 use topcoat::{
-    context::CxTestBuilder,
+    context::test::CxTestBuilder,
     mail::{MailConfig, MemoryTransport, Mailbox, mail, send},
 };
